+++ Erebus/LuaCamera.cpp
@@ -32,12 +32,12 @@ namespace LuaCamera {
 		int transformIndex;
 		float distance, angle, xoffset, yoffset, fov;
 		
-		angle = lua_tonumber(lua, -1);
-		distance = lua_tonumber(lua, -2);
-		xoffset = lua_tonumber(lua, -3);
-		yoffset = lua_tonumber(lua, -4);
-		transformIndex = lua_tointeger(lua, -5);
-		fov = lua_tonumber(lua, -6);
+		angle = (float)lua_tonumber(lua, -1);
+		distance = (float)lua_tonumber(lua, -2);
+		xoffset = (float)lua_tonumber(lua, -3);
+		yoffset = (float)lua_tonumber(lua, -4);
+		transformIndex = (int)lua_tointeger(lua, -5);
+		fov = (float)lua_tonumber(lua, -6);
 		pos = g_transforms[transformIndex].getPos();
 		dir = g_transforms[transformIndex].getLookAt();
 
@@ -50,12 +50,12 @@ namespace LuaCamera {
 
 		glm::vec3 campos, lookpos;
 
-		campos.x = lua_tonumber(lua, -6);
-		campos.y = lua_tonumber(lua, -5);
-		campos.z = lua_tonumber(lua, -4);
-		lookpos.x = lua_tonumber(lua, -3);
-		lookpos.y = lua_tonumber(lua, -2);
-		lookpos.z = lua_tonumber(lua, -1);
+		campos.x = (float)lua_tonumber(lua, -6);
+		campos.y = (float)lua_tonumber(lua, -5);
+		campos.z = (float)lua_tonumber(lua, -4);
+		lookpos.x = (float)lua_tonumber(lua, -3);
+		lookpos.y = (float)lua_tonumber(lua, -2);
+		lookpos.z = (float)lua_tonumber(lua, -1);
 
 		g_camera->setCamera(campos, lookpos);
 
@@ -64,7 +64,7 @@ namespace LuaCamera {
 
 	int getPos(lua_State* lua) {
 
-		int index = lua_tointeger(lua, 1);
+		int index = (int)lua_tointeger(lua, 1);
 
 		glm::vec3 position = g_camera->getPosition();
 
@@ -82,13 +82,13 @@ namespace LuaCamera {
 	}
 	int setHeight(lua_State * lua)
 	{
-		float height = lua_tonumber(lua, -1);
+		float height = (float)lua_tonumber(lua, -1);
 		g_camera->setHeight(height);
 		return 1;
 	}
 	int getDirection(lua_State * lua)
 	{
-		int index = lua_tointeger(lua, 1);
+		int index = (int)lua_tointeger(lua, 1);
 
 		glm::vec3 position = g_camera->getDirection();
